+++ examples/historical/timeseries_get_range_to_file.cpp
@@ -1,8 +1,3 @@
-#include <chrono>
-#include <ctime>
-#include <iomanip>
-#include <iostream>  // setw
-
 #include "databento/constants.hpp"
 #include "databento/datetime.hpp"
 #include "databento/dbn_file_store.hpp"
@@ -10,17 +5,19 @@
 #include "databento/historical.hpp"
 #include "databento/record.hpp"
 
+namespace db = databento;
+
 int main() {
-  auto client = databento::HistoricalBuilder{}.SetKeyFromEnv().Build();
+  auto client = db::Historical::Builder().SetKeyFromEnv().Build();
   const auto limit = 1000;
-  databento::DbnFileStore dbn_file_store = client.TimeseriesGetRangeToFile(
-      databento::dataset::kGlbxMdp3, {"2022-10-03T00:00", "2022-10-04T00:00"}, {"ESZ2"},
-      databento::Schema::Ohlcv1M, databento::SType::RawSymbol,
-      databento::SType::InstrumentId, limit, "ESZ2-ohlcv1m-20201003-20201004.dbn.zst");
-  dbn_file_store.Replay([](const databento::Record record) {
-    const auto& ohlcv_bar = record.Get<databento::OhlcvMsg>();
+  db::DbnFileStore dbn_file_store = client.TimeseriesGetRangeToFile(
+      db::dataset::kGlbxMdp3, {"2022-10-03T00:00", "2022-10-04T00:00"}, {"ESZ2"},
+      db::Schema::Ohlcv1M, db::SType::RawSymbol, db::SType::InstrumentId, limit,
+      "ESZ2-ohlcv1m-20201003-20201004.dbn.zst");
+  dbn_file_store.Replay([](const db::Record record) {
+    const auto& ohlcv_bar = record.Get<db::OhlcvMsg>();
     std::cout << ohlcv_bar << '\n';
-    return databento::KeepGoing::Continue;
+    return db::KeepGoing::Continue;
   });
   return 0;
 }