+++ src/historical.cpp
@@ -121,6 +121,10 @@ void TryCreateDir(const std::filesystem::path& dir_name) {
 }
 }  // namespace
 
+databento::HistoricalBuilder Historical::Builder() {
+  return databento::HistoricalBuilder{};
+}
+
 Historical::Historical(ILogReceiver* log_receiver, std::string key,
                        HistoricalGateway gateway)
     : log_receiver_{log_receiver},
@@ -227,7 +231,7 @@ databento::BatchJob Historical::BatchSubmitJob(const httplib::Params& params) {
 
 std::vector<databento::BatchJob> Historical::BatchListJobs() {
   static const std::vector<JobState> kDefaultStates = {
-      JobState::Received, JobState::Queued, JobState::Processing, JobState::Done};
+      JobState::Queued, JobState::Processing, JobState::Done};
   return this->BatchListJobs(kDefaultStates, UnixNanos{});
 }
 std::vector<databento::BatchJob> Historical::BatchListJobs(